+++ src/main/java/com/ssafy/bango/global/batch/job/v1/FacilityBatchConfiguration.java
@@ -9,12 +9,11 @@ import com.ssafy.bango.global.batch.chunk.facility.v1.FacilityApiItemProcessor;
 import com.ssafy.bango.global.batch.chunk.facility.v1.FacilityApiItemReader;
 import com.ssafy.bango.global.batch.chunk.facility.v1.FacilityApiItemWriter;
 import com.ssafy.bango.global.batch.job.listener.CustomStepExecutionListener;
-import com.ssafy.bango.global.batch.tasklet.FacilityDeleteDataTasklet;
+import com.ssafy.bango.global.batch.tasklet.v1.FacilityDeleteDataTasklet;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.Step;
-import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
 import org.springframework.batch.core.configuration.annotation.StepScope;
 import org.springframework.batch.core.job.builder.JobBuilder;
 import org.springframework.batch.core.repository.JobRepository;
